Bug#678311: apt-listchanges: debconf passthrough frontend for use with packagekit
Package: apt-listchanges
Version: 2.85.10
Severity: normal
Tags: patch
X-Debbugs-Cc: packagekit@packages.debian.org, debconf@packages.debian.org

Hi,

when using packagekit, which seems to be the way gnome is going for
graphical package management nowadays, the apt-listchanges output
doesn't go anywhere, and the gpk window just seems to hang.
In order to fix that, it'd be nice if apt-listchanges could send its
output using the debconf protocol so it could show up in the packagekit
frontend.  The below patch is a proof of concept to that effect.  pk
would have to set APT_LISTCHANGES_FRONTEND=debconf for this to work (it
already sets APT_LISTBUGS_FRONTEND=none).  packagekit and debconf
maintainers cc:ed in case they have comments.

I'm not too happy about the tempfile thing instead of just using
db.subst to change the contents of the note, but I couldn't figure out
how to SUBST something with newlines (it didn't seem to work when I
tested with the dialog frontend, anyway).

Cheers,
Julien